Move pixelsToPNG into RenderTargetRecorder
